move common iast controller methods to a single shared file.
