+++ tracer/src/Datadog.Trace/IAST/StackWalker.cs
@@ -24,12 +24,13 @@ internal static class StackWalker
         "Oracle.DataAccess",
         "Oracle.ManagedDataAccess",
         "System.Data",
+        "System.Data.Common",
         "System.Security.Cryptography",
         "System.Security.Cryptography.Algorithms",
         "System.Security.Cryptography.Csp",
         "System.Security.Cryptography.Primitives",
         "System.Data.SqlClient",
-        "System.Data.SQLite",
+        "System.Data.SQLite"
     };
 
     private const int DefaultSkipFrames = 2;
