+++ tracer/test/test-applications/security/aspnet/Samples.Security.AspNetMvc5/Samples.Security.AspNetMvc5.csproj
@@ -161,6 +161,7 @@
     <Compile Include="Controllers\PostsController.cs" />
     <Compile Include="Controllers\RenderController.cs" />
     <Compile Include="Data\DatabaseHelper.cs" />
+	<Compile Include="..\..\Samples.Security.AspNetCore5\Data\IastControllerHelper.cs" Link="Data\IastControllerHelper.cs" />
     <Compile Include="Global.asax.cs">
       <DependentUpon>Global.asax</DependentUpon>
     </Compile>
